refactor(sidekick): split `model.go`
This file was too large to navigate. Generally, split each type and its methods
to its own file.
