+++ internal/sidekick/api/model.go
@@ -15,152 +15,6 @@
 // Package api defines the data model representing a parsed API surface.
 package api
 
-import (
-	"fmt"
-	"iter"
-	"maps"
-	"slices"
-	"strings"
-)
-
-// Typez represent different field types that may be found in messages.
-type Typez int
-
-// These are the different field types as defined in
-// descriptorpb.FieldDescriptorProto_Type.
-const (
-	TypezUndefined Typez = 0
-	TypezDouble    Typez = 1
-	TypezFloat     Typez = 2
-	TypezInt64     Typez = 3
-	TypezUint64    Typez = 4
-	TypezInt32     Typez = 5
-	TypezFixed64   Typez = 6
-	TypezFixed32   Typez = 7
-	TypezBool      Typez = 8
-	TypezString    Typez = 9
-	TypezGroup     Typez = 10
-	TypezMessage   Typez = 11
-	TypezBytes     Typez = 12
-	TypezUint32    Typez = 13
-	TypezEnum      Typez = 14
-	TypezSfixed32  Typez = 15
-	TypezSfixed64  Typez = 16
-	TypezSint32    Typez = 17
-	TypezSint64    Typez = 18
-)
-
-var typezName = [...]string{
-	"UNDEFINED",
-	"DOUBLE",
-	"FLOAT",
-	"INT64",
-	"UINT64",
-	"INT32",
-	"FIXED64",
-	"FIXED32",
-	"BOOL",
-	"STRING",
-	"GROUP",
-	"MESSAGE",
-	"BYTES",
-	"UINT32",
-	"ENUM",
-	"SFIXED32",
-	"SFIXED64",
-	"SINT32",
-	"SINT64",
-}
-
-// String returns the symbolic name for the Typez.
-func (t Typez) String() string {
-	if t < 0 || int(t) >= len(typezName) {
-		return fmt.Sprintf("Typez(%d)", t)
-	}
-	return typezName[t]
-}
-
-// FieldBehavior represents annotations for how the code generator handles a
-// field.
-//
-// Regardless of the underlying data type and whether it is required or optional
-// on the wire, some fields must be present for requests to succeed. Or may not
-// be included in a request.
-type FieldBehavior int
-
-const (
-	// FieldBehaviorUnspecified is the default, unspecified field behavior.
-	FieldBehaviorUnspecified FieldBehavior = iota
-
-	// FieldBehaviorOptional specifically denotes a field as optional.
-	//
-	// While Google Cloud uses proto3, where fields are either optional or have
-	// a default value, this may be specified for emphasis.
-	FieldBehaviorOptional
-
-	// FieldBehaviorRequired denotes a field as required.
-	//
-	// This indicates that the field **must** be provided as part of the request,
-	// and failure to do so will cause an error (usually `INVALID_ARGUMENT`).
-	//
-	// Code generators may change the generated types to include this field as a
-	// parameter necessary to construct the request.
-	FieldBehaviorRequired
-
-	// FieldBehaviorOutputOnly denotes a field as output only.
-	//
-	// Some messages (and their fields) are used in both requests and responses.
-	// This indicates that the field is provided in responses, but including the
-	// field in a request does nothing (the server *must* ignore it and
-	// *must not* throw an error as a result of the field's presence).
-	//
-	// Code generators that use different builders for "the message as part of a
-	// request" vs. "the standalone message" may omit this field in the former.
-	FieldBehaviorOutputOnly
-
-	// FieldBehaviorInputOnly denotes a field as input only.
-	//
-	// This indicates that the field is provided in requests, and the
-	// corresponding field is not included in output.
-	FieldBehaviorInputOnly
-
-	// FieldBehaviorImmutable denotes a field as immutable.
-	//
-	// This indicates that the field may be set once in a request to create a
-	// resource, but may not be changed thereafter.
-	FieldBehaviorImmutable
-
-	// FieldBehaviorUnorderedList denotes that a (repeated) field is an unordered list.
-	//
-	// This indicates that the service may provide the elements of the list
-	// in any arbitrary  order, rather than the order the user originally
-	// provided. Additionally, the list's order may or may not be stable.
-	FieldBehaviorUnorderedList
-
-	// FieldBehaviorUnorderedNonEmptyDefault denotes that this field returns a non-empty default value if not set.
-	//
-	// This indicates that if the user provides the empty value in a request,
-	// a non-empty value will be returned. The user will not be aware of what
-	// non-empty value to expect.
-	FieldBehaviorUnorderedNonEmptyDefault
-
-	// FieldBehaviorIdentifier denotes that the field in a resource (a message annotated with
-	// google.api.resource) is used in the resource name to uniquely identify the
-	// resource.
-	//
-	// For AIP-compliant APIs, this should only be applied to the
-	// `name` field on the resource.
-	//
-	// This behavior should not be applied to references to other resources within
-	// the message.
-	//
-	// The identifier field of resources often have different field behavior
-	// depending on the request it is embedded in (e.g. for Create methods name
-	// is optional and unused, while for Update methods it is required). Instead
-	// of method-specific annotations, only `IDENTIFIER` is required.
-	FieldBehaviorIdentifier
-)
-
 const (
 	// ReservedPackageName is a package name reserved for maps and other
 	// synthetic messages that do not exist in the input specification.
@@ -168,415 +22,21 @@ const (
 	// We need a place to put these in the data model without conflicts with the
 	// input data model. This symbol is unused in all the IDLs we support.
 	ReservedPackageName = "$"
-)
 
-// API represents and API surface.
-type API struct {
-	// Name of the API (e.g. secretmanager).
-	Name string
-	// Name of the package name in the source specification format. For Protobuf
-	// this may be `google.cloud.secretmanager.v1`.
-	PackageName string
-	// The API Title (e.g. "Secret Manager API" or "Cloud Spanner API").
-	Title string
-	// The API Description.
-	Description string
-	// The API Revision. In discovery-based services this is the "revision"
-	// attribute.
-	Revision string
-	// Services are a collection of services that make up the API.
-	Services []*Service
-	// Messages are a collection of messages used to process request and
-	// responses in the API.
-	Messages []*Message
-	// Enums
-	Enums []*Enum
-	// ResourceDefinitions contains the data from the `google.api.resource_definition` annotation.
-	ResourceDefinitions []*Resource
-	// QuickstartService is the service that will be used to generate the quickstart sample
-	// at the package level.
-	QuickstartService *Service
-	// Language specific annotations.
-	Codec any
-
-	// serviceByID returns a service that is associated with the API.
-	serviceByID map[string]*Service
-	// methodByID returns a method that is associated with the API.
-	methodByID map[string]*Method
-	// messageByID returns a message that is associated with the API.
-	messageByID map[string]*Message
-	// enumByID returns a message that is associated with the API.
-	enumByID map[string]*Enum
-	// resourceByType returns a resource that is associated with the API.
-	resourceByType map[string]*Resource
-}
-
-// ModelOverride holds configuration overrides for an API model.
-type ModelOverride struct {
-	Name        string
-	Title       string
-	Description string
-	IncludedIDs []string
-	SkippedIDs  []string
-}
-
-// HasMessages returns true if the API contains messages (most do).
-//
-// This is useful in the mustache templates to skip code that only makes sense
-// when per-message code follows.
-func (api *API) HasMessages() bool {
-	return len(api.Messages) != 0
-}
-
-// ModelCodec returns the Codec field with an alternative name.
-//
-// In some mustache templates we want to access the annotations for the
-// enclosing model. In mustache you can get a field from an enclosing context
-// *if* the name is unique.
-func (a *API) ModelCodec() any {
-	return a.Codec
-}
-
-// Service returns a service that is associated with the API.
-func (a *API) Service(id string) *Service {
-	return a.serviceByID[id]
-}
-
-// AllServices returns an iterator over the services in the API.
-func (a *API) AllServices() iter.Seq[*Service] {
-	return maps.Values(a.serviceByID)
-}
-
-// AddService adds a service to the API.
-func (a *API) AddService(s *Service) {
-	if a.serviceByID == nil {
-		a.serviceByID = make(map[string]*Service)
-	}
-	a.serviceByID[s.ID] = s
-}
-
-// Method returns a method that is associated with the API.
-func (a *API) Method(id string) *Method {
-	return a.methodByID[id]
-}
-
-// AllMethods returns an iterator over the methods in the API.
-func (a *API) AllMethods() iter.Seq[*Method] {
-	return maps.Values(a.methodByID)
-}
-
-// AddMethod adds a method to the API.
-func (a *API) AddMethod(m *Method) {
-	if a.methodByID == nil {
-		a.methodByID = make(map[string]*Method)
-	}
-	a.methodByID[m.ID] = m
-}
-
-// Message returns a message that is associated with the API.
-func (a *API) Message(id string) *Message {
-	return a.messageByID[id]
-}
-
-// AllMessages returns an iterator over the messages in the API.
-func (a *API) AllMessages() iter.Seq[*Message] {
-	return maps.Values(a.messageByID)
-}
-
-// AddMessage adds a message to the API.
-func (a *API) AddMessage(m *Message) {
-	if a.messageByID == nil {
-		a.messageByID = make(map[string]*Message)
-	}
-	a.messageByID[m.ID] = m
-}
-
-// Enum returns a message that is associated with the API.
-func (a *API) Enum(id string) *Enum {
-	return a.enumByID[id]
-}
-
-// AllEnums returns an iterator over the enums in the API.
-func (a *API) AllEnums() iter.Seq[*Enum] {
-	return maps.Values(a.enumByID)
-}
-
-// AddEnum adds an enum to the API.
-func (a *API) AddEnum(e *Enum) {
-	if a.enumByID == nil {
-		a.enumByID = make(map[string]*Enum)
-	}
-	a.enumByID[e.ID] = e
-}
-
-// Resource returns a resource that is associated with the API.
-func (a *API) Resource(typ string) *Resource {
-	return a.resourceByType[typ]
-}
-
-// AllResources returns an iterator over the resources in the API.
-func (a *API) AllResources() iter.Seq[*Resource] {
-	return maps.Values(a.resourceByType)
-}
-
-// AddResource adds a resource to the API.
-func (a *API) AddResource(r *Resource) {
-	if a.resourceByType == nil {
-		a.resourceByType = make(map[string]*Resource)
-	}
-	a.resourceByType[r.Type] = r
-}
-
-// Service represents a service in an API.
-type Service struct {
-	// Documentation for the service.
-	Documentation string
-	// Name of the attribute.
-	Name string
-	// ID is a unique identifier.
-	ID string
-	// Some source specifications allow marking services as deprecated.
-	Deprecated bool
-	// Methods associated with the Service.
-	Methods []*Method
-	// DefaultHost fragment of a URL.
-	DefaultHost string
-	// The Protobuf package this service belongs to.
-	Package string
-
-	// The model this service belongs to, mustache templates use this field to
-	// navigate the data structure.
-	Model *API
-	// QuickstartMethod is the method that will be used to generate the quickstart sample
-	// for this service.
-	QuickstartMethod *Method
-	// Language specific annotations.
-	Codec any
-}
-
-// HasClientSideStreaming returns true if the service contains any methods
-// that support client-side streaming.
-func (s *Service) HasClientSideStreaming() bool {
-	return slices.ContainsFunc(s.Methods, func(m *Method) bool {
-		return m.ClientSideStreaming
-	})
-}
-
-// Method defines a RPC belonging to a Service.
-type Method struct {
-	// Documentation is the documentation for the method.
-	Documentation string
-	// Name is the name of the attribute.
-	Name string
-	// ID is a unique identifier.
-	ID string
-	// Deprecated is true if the method is deprecated.
-	Deprecated bool
-	// InputTypeID is the ID of the input type for the Method.
-	InputTypeID string
-	// InputType is the input to the Method.
-	InputType *Message
-	// OutputTypeID is the ID of the output type for the Method.
-	OutputTypeID string
-	// OutputType is the output of the Method.
-	OutputType *Message
-	// ReturnsEmpty is true if the method returns nothing.
+	// StandardFieldNameForResourceRef is the standard name for the resource
+	// field.
 	//
-	// Protobuf uses the well-known type `google.protobuf.Empty` message to
-	// represent this.
-	//
-	// OpenAPIv3 uses a missing content field:
-	//   https://swagger.io/docs/specification/v3_0/describing-responses/#empty-response-body
-	ReturnsEmpty bool
-	// PathInfo contains information about the HTTP request.
-	PathInfo *PathInfo
-	// Pagination holds the `page_token` field if the method conforms to the
-	// standard defined by [AIP-4233](https://google.aip.dev/client-libraries/4233).
-	Pagination *Field
-	// ClientSideStreaming is true if the method supports client-side streaming.
-	ClientSideStreaming bool
-	// ServerSideStreaming is true if the method supports server-side streaming.
-	ServerSideStreaming bool
-	// IsLroPoller is true if the method is an LRO poller.
-	IsLroPoller bool
-	// OperationInfo contains information for methods returning long-running operations.
-	OperationInfo *OperationInfo
-	// DiscoveryLro has a value if this is a discovery-style long-running operation.
-	DiscoveryLro *DiscoveryLro
-	// Routing contains the routing annotations, if any.
-	Routing []*RoutingInfo
-	// AutoPopulated contains the auto-populated (request_id) field, if any, as defined in
-	// [AIP-4235](https://google.aip.dev/client-libraries/4235)
-	//
-	// The field must be eligible for auto-population, and be listed in the
-	// `google.api.MethodSettings.auto_populated_fields` entry in
-	// `google.api.Publishing.method_settings` in the service config file.
-	AutoPopulated []*Field
-	// APIVersion contains the interface-based-versioning version.
-	//
-	// If this is empty, then the method does not have a version annotation.
-	APIVersion string
-	// Model is the model this method belongs to, mustache templates use this field to
-	// navigate the data structure.
-	Model *API
-	// Service is the service this method belongs to, mustache templates use this field to
-	// navigate the data structure.
-	Service *Service
-	// `SourceService` is the original service this method belongs to. For most
-	// methods `SourceService` and `Service` are the same. For mixins, the
-	// source service is the mixin, such as longrunning.Operations.
-	SourceService *Service
-	// `SourceServiceID` is the original service ID for this method.
-	SourceServiceID string
-	// IsSimple is true if the method is not a streaming, pagination or LRO method.
-	IsSimple bool
-	// IsLRO is true if the method is a long-running operation.
-	IsLRO bool
-	// LongRunningResponseType is the response type of the long-running operation.
-	LongRunningResponseType *Message
-	// LongRunningReturnsEmpty is true if the long-running operation returns an empty response.
-	LongRunningReturnsEmpty bool
-	// IsList is true if the method is a list operation.
-	IsList bool
-	// IsStreaming is true if the method is client-side or server-side streaming.
-	IsStreaming bool
-	// IsAIPStandard is true if the method is one of the AIP standard methods.
-	IsAIPStandard bool
-	// IsAIPStandardGet is true if the method is an AIP standard get method.
-	IsAIPStandardGet bool
-	// IsAIPStandardDelete is true if the method is an AIP standard delete method.
-	IsAIPStandardDelete bool
-	// IsAIPStandardUndelete is true if the method is an AIP standard undelete method.
-	IsAIPStandardUndelete bool
-	// IsAIPStandardCreate is true if the method is an AIP standard create method.
-	IsAIPStandardCreate bool
-	// IsAIPStandardUpdate is true if the method is an AIP standard update method.
-	IsAIPStandardUpdate bool
-	// IsAIPStandardList is true if the method is an AIP standard list method.
-	IsAIPStandardList bool
-	// SampleInfo may contain sample generation information for this method,
-	// usually if it is an AIP conforming metho.
-	SampleInfo *SampleInfo
-	// Signatures defines alternative signatures (overloads) for the method.
-	Signatures []*MethodSignature
-	// Codec contains language specific annotations.
-	Codec any
-}
-
-// RoutingCombos returns all combinations of routing parameters.
-//
-// The routing info is stored as a map from the key to a list of the variants.
-// e.g.:
-//
-// ```
-//
-//	{
-//	  a: [va1, va2, va3],
-//	  b: [vb1, vb2]
-//	  c: [vc1]
-//	}
-//
-// ```
-//
-// We reorganize each kv pair into a list of pairs. e.g.:
-//
-// ```
-// [
-//
-//	[(a, va1), (a, va2), (a, va3)],
-//	[(b, vb1), (b, vb2)],
-//	[(c, vc1)],
-//
-// ]
-// ```
-//
-// Then we take a Cartesian product of that list to find all the combinations.
-// e.g.:
-//
-// ```
-// [
-//
-//	[(a, va1), (b, vb1), (c, vc1)],
-//	[(a, va1), (b, vb2), (c, vc1)],
-//	[(a, va2), (b, vb1), (c, vc1)],
-//	[(a, va2), (b, vb2), (c, vc1)],
-//	[(a, va3), (b, vb1), (c, vc1)],
-//	[(a, va3), (b, vb2), (c, vc1)],
-//
-// ]
-// ```.
-func (m *Method) RoutingCombos() []*RoutingInfoCombo {
-	combos := []*RoutingInfoCombo{
-		{},
-	}
-	for _, info := range m.Routing {
-		next := []*RoutingInfoCombo{}
-		for _, c := range combos {
-			for _, v := range info.Variants {
-				next = append(next, &RoutingInfoCombo{
-					Items: append(c.Items, &RoutingInfoComboItem{
-						Name:    info.Name,
-						Variant: v,
-					}),
-				})
-			}
-		}
-		combos = next
-	}
-	return combos
-}
-
-// RoutingInfoCombo represents a single combination of routing parameters.
-type RoutingInfoCombo struct {
-	Items []*RoutingInfoComboItem
-}
-
-// RoutingInfoComboItem represents a single item in a RoutingInfoCombo.
-type RoutingInfoComboItem struct {
-	Name    string
-	Variant *RoutingInfoVariant
-}
-
-// HasRouting returns true if the method has routing information.
-func (m *Method) HasRouting() bool {
-	return len(m.Routing) != 0
-}
-
-// HasAutoPopulatedFields returns true if the method has auto-populated fields.
-func (m *Method) HasAutoPopulatedFields() bool {
-	return len(m.AutoPopulated) != 0
-}
-
-// SampleInfo contains sample generation information for a single method,
-// usually if it is an AIP conforming method.
-type SampleInfo struct {
-	// ResourceNameField is the field containing the resource name or parent resource name.
-	ResourceNameField *Field
-	// ResourceIDField is the field containing the resource ID, usually present in Create methods.
-	ResourceIDField *Field
-	// MessageField is the field containing the message body to be created or updated.
-	MessageField *Field
-	// UpdateMaskField is the field containing the update mask, present in Update methods.
-	UpdateMaskField *Field
-	// IsRequestResourceName is true if the main resource name associated to this sample,
-	// i.e. the name of the resource manipulated by the RPC, is a field of the request
-	// object itself, e.g. for Delete or Get operations.
-	IsRequestResourceName bool
-	// IsMessageResourceName is true if the main resource name associated to this sample,
-	// i.e. the name of the resource manipulated by the RPC, is a field of the message body of
-	// the request object, e.g. for Update operations.
-	IsMessageResourceName bool
-	// Codec is a placeholder to put language specific annotations.
-	Codec any
-}
-
-const (
-	// StandardFieldNameForResourceRef is the standard name for resource references
-	// to the resource being operated on by standard methods as defined by AIPs.
+	// That is, the field that contains the resource name that an RPC operates
+	// on. This is one of the AIP requirements / recommendations for standard
+	// RPCs.
 	StandardFieldNameForResourceRef = "name"
 
-	// StandardFieldNameForParentResourceRef is the standard name for resource references
-	// to the child resource being operated on by standard methods as defined by AIPs.
+	// StandardFieldNameForParentResourceRef is the standard name for the child
+	// resource field.
+	//
+	// That is, the field that contains the child resource name that an RPC
+	// operates on. This is one of the AIP requirements / recommendations for
+	// standard RPCs.
 	StandardFieldNameForParentResourceRef = "parent"
 
 	// GenericResourceType is a special resource type that may be used by resource references
@@ -587,646 +47,3 @@ const (
 	// in update operations as defined by AIP-134.
 	StandardFieldNameForUpdateMask = "update_mask"
 )
-
-// PathInfo contains normalized request path information.
-type PathInfo struct {
-	// The list of bindings, including the top-level binding.
-	Bindings []*PathBinding
-	// Body is the name of the field that should be used as the body of the
-	// request.
-	//
-	// This is a string that may be "*" which indicates that the entire request
-	// should be used as the body.
-	//
-	// If this is empty then the body is not used.
-	BodyFieldPath string
-	// Language specific annotations.
-	Codec any
-}
-
-// PathBinding is a binding of a path to a method.
-type PathBinding struct {
-	// HTTP Verb.
-	//
-	// This is one of:
-	// - GET
-	// - POST
-	// - PUT
-	// - DELETE
-	// - PATCH
-	Verb string
-	// The path broken by components.
-	PathTemplate *PathTemplate
-	// Query parameter fields.
-	QueryParameters map[string]bool
-	// TargetResource contains the results of the resource name identification.
-	// This helps identify which resource this path is likely targeting.
-	TargetResource *TargetResource
-	// Language specific annotations.
-	Codec any
-}
-
-// OperationInfo contains normalized long running operation info.
-type OperationInfo struct {
-	// The metadata type. If there is no metadata, this is set to
-	// `.google.protobuf.Empty`.
-	MetadataTypeID string
-	// The result type. This is the expected type when the long running
-	// operation completes successfully.
-	ResponseTypeID string
-	// The method.
-	Method *Method
-	// Language specific annotations.
-	Codec any
-}
-
-// DiscoveryLro contains old-style long-running operation descriptors.
-type DiscoveryLro struct {
-	// The path parameters required by the polling operation.
-	PollingPathParameters []string
-	// Language specific annotations.
-	Codec any
-}
-
-// RoutingInfo contains normalized routing info.
-//
-// The routing information format is documented in:
-//
-// https://google.aip.dev/client-libraries/4222
-//
-// At a high level, it consists of a field name (from the request) that is used
-// to match a certain path template. If the value of the field matches the
-// template, the matching portion is added to `x-goog-request-params`.
-//
-// An empty `Name` field is used as the special marker to cover this case in
-// AIP-4222:
-//
-//	An empty google.api.routing annotation is acceptable. It means that no
-//	routing headers should be generated for the RPC, when they otherwise
-//	would be e.g. implicitly from the google.api.http annotation.
-type RoutingInfo struct {
-	// The name in `x-goog-request-params`.
-	Name string
-	// Group the possible variants for the given name.
-	//
-	// The variants are parsed into the reverse order of definition. AIP-4222
-	// declares:
-	//
-	//   In cases when multiple routing parameters have the same resource ID
-	//   path segment name, thus referencing the same header key, the
-	//   "last one wins" rule is used to determine which value to send.
-	//
-	// Reversing the order allows us to implement "the first match wins". That
-	// is easier and more efficient in most languages.
-	Variants []*RoutingInfoVariant
-}
-
-// RoutingInfoVariant represents the routing information stripped of its name.
-type RoutingInfoVariant struct {
-	// The sequence of field names accessed to get the routing information.
-	FieldPath []string
-	// A path template that must match the beginning of the field value.
-	Prefix RoutingPathSpec
-	// A path template that, if matching, is used in the `x-goog-request-params`.
-	Matching RoutingPathSpec
-	// A path template that must match the end of the field value.
-	Suffix RoutingPathSpec
-	// Language specific information
-	Codec any
-}
-
-// FieldName returns the field path as a string.
-func (v *RoutingInfoVariant) FieldName() string {
-	return strings.Join(v.FieldPath, ".")
-}
-
-// TemplateAsString returns the template as a string.
-func (v *RoutingInfoVariant) TemplateAsString() string {
-	var full []string
-	full = append(full, v.Prefix.Segments...)
-	full = append(full, v.Matching.Segments...)
-	full = append(full, v.Suffix.Segments...)
-	return strings.Join(full, "/")
-}
-
-// RoutingPathSpec is a specification for a routing path.
-type RoutingPathSpec struct {
-	// A sequence of matching segments.
-	//
-	// A template like `projects/*/location/*/**` maps to
-	// `["projects", "*", "locations", "*", "**"]`.
-	Segments []string
-}
-
-const (
-	// SingleSegmentWildcard is a special routing path segment which indicates
-	// "match anything that does not include a `/`".
-	SingleSegmentWildcard = "*"
-
-	// MultiSegmentWildcard is a special routing path segment which indicates
-	// "match anything including `/`".
-	MultiSegmentWildcard = "**"
-)
-
-// PathTemplate is a template for a path.
-type PathTemplate struct {
-	Segments []PathSegment
-	Verb     string
-}
-
-// FlatPath returns a simplified representation of the path template as a string.
-//
-// In the context of discovery LROs it is useful to get the path template as a
-// simplified string, such as "compute/v1/projects/{project}/zones/{zone}/instances".
-// The path can be matched against LRO prefixes and then mapped to the correct
-// poller RPC.
-func (template *PathTemplate) FlatPath() string {
-	var buffer strings.Builder
-	sep := ""
-	for _, segment := range template.Segments {
-		buffer.WriteString(sep)
-		if segment.Literal != "" {
-			buffer.WriteString(segment.Literal)
-		} else if segment.Variable != nil {
-			fmt.Fprintf(&buffer, "{%s}", strings.Join(segment.Variable.FieldPath, "."))
-		}
-		sep = "/"
-	}
-	return buffer.String()
-}
-
-// PathSegment is a segment of a path.
-type PathSegment struct {
-	Literal  string
-	Variable *PathVariable
-}
-
-// PathVariable is a variable in a path.
-type PathVariable struct {
-	FieldPath []string
-	Segments  []string
-	// Allow characters defined as `reserved` by RFC-6570 1.5 to pass through without
-	// percent encoding. See RFC-6570 1.2 for examples.
-	AllowReserved bool
-}
-
-// NewPathVariable creates a new path variable.
-func NewPathVariable(fields ...string) *PathVariable {
-	return &PathVariable{FieldPath: fields}
-}
-
-// WithLiteral adds a literal to the path template.
-func (p *PathTemplate) WithLiteral(l string) *PathTemplate {
-	p.Segments = append(p.Segments, PathSegment{Literal: l})
-	return p
-}
-
-// WithVariable adds a variable to the path template.
-func (p *PathTemplate) WithVariable(v *PathVariable) *PathTemplate {
-	p.Segments = append(p.Segments, PathSegment{Variable: v})
-	return p
-}
-
-// WithVariableNamed adds a variable with the given name to the path template.
-func (p *PathTemplate) WithVariableNamed(fields ...string) *PathTemplate {
-	v := PathVariable{FieldPath: fields}
-	p.Segments = append(p.Segments, PathSegment{Variable: v.WithMatch()})
-	return p
-}
-
-// WithVerb adds a verb to the path template.
-func (p *PathTemplate) WithVerb(v string) *PathTemplate {
-	p.Verb = v
-	return p
-}
-
-// WithLiteral adds a literal to the path variable.
-func (v *PathVariable) WithLiteral(l string) *PathVariable {
-	v.Segments = append(v.Segments, l)
-	return v
-}
-
-// WithMatchRecursive adds a recursive match to the path variable.
-func (v *PathVariable) WithMatchRecursive() *PathVariable {
-	v.Segments = append(v.Segments, MultiSegmentWildcard)
-	return v
-}
-
-// WithMatch adds a match to the path variable.
-func (v *PathVariable) WithMatch() *PathVariable {
-	v.Segments = append(v.Segments, SingleSegmentWildcard)
-	return v
-}
-
-// WithAllowReserved marks the variable as allowing reserved characters to remain unescaped.
-func (v *PathVariable) WithAllowReserved() *PathVariable {
-	v.AllowReserved = true
-	return v
-}
-
-// WithLiteral adds a literal to the path segment.
-func (s *PathSegment) WithLiteral(l string) *PathSegment {
-	s.Literal = l
-	return s
-}
-
-// WithVariable adds a variable to the path segment.
-func (s *PathSegment) WithVariable(v *PathVariable) *PathSegment {
-	s.Variable = v
-	return s
-}
-
-// Message defines a message used in request/response handling.
-type Message struct {
-	// Documentation for the message.
-	Documentation string
-	// Name of the attribute.
-	Name string
-	// ID is a unique identifier.
-	ID string
-	// Some source specifications allow marking messages as deprecated.
-	Deprecated bool
-	// Fields associated with the Message.
-	Fields []*Field
-	// If true, this is a synthetic request message.
-	//
-	// These messages are created by sidekick when parsing Discovery docs and
-	// OpenAPI specifications. The query and request parameters for each method
-	// are grouped into a synthetic message.
-	SyntheticRequest bool
-	// If true, this message is a placeholder / doppelganger for a `api.Service`.
-	//
-	// These messages are created by sidekick when parsing Discovery docs and
-	// OpenAPI specifications. All the synthetic messages for a service need to
-	// be grouped under a unique namespace to avoid clashes with similar
-	// synthetic messages in other services. Sidekick creates a placeholder
-	// message that represents "the service".
-	//
-	// That is, `service1` and `service2` may both have a synthetic `getRequest`
-	// message, with different attributes. We need these to be different
-	// messages, with different names. So we create a different parent message
-	// for each.
-	ServicePlaceholder bool
-	// Enums associated with the Message.
-	Enums []*Enum
-	// Messages associated with the Message. In protobuf these are referred to as
-	// nested messages.
-	Messages []*Message
-	// OneOfs associated with the Message.
-	OneOfs []*OneOf
-	// Parent returns the ancestor of this message, if any.
-	Parent *Message
-	// The Protobuf package this message belongs to.
-	Package string
-	IsMap   bool
-	// Indicates that this Message is returned by a standard
-	// List RPC and conforms to [AIP-4233](https://google.aip.dev/client-libraries/4233).
-	Pagination *PaginationInfo
-	// Resource contains the data from the `google.api.resource` annotation.
-	Resource *Resource
-	// Language specific annotations.
-	Codec any
-}
-
-// HasFields returns true if the message has fields.
-func (m *Message) HasFields() bool {
-	return len(m.Fields) != 0
-}
-
-// Enum defines a message used in request/response handling.
-type Enum struct {
-	// Documentation for the message.
-	Documentation string
-	// Name of the attribute.
-	Name string
-	// ID is a unique identifier.
-	ID string
-	// Some source specifications allow marking enums as deprecated.
-	Deprecated bool
-	// Values associated with the Enum.
-	Values []*EnumValue
-	// The unique integer values, some enums have multiple aliases for the
-	// same number (e.g. `enum X { a = 0, b = 0, c = 1 }`).
-	UniqueNumberValues []*EnumValue
-	// ValuesForExamples contains a subset of values suitable for use in generated samples.
-	// e.g. non-deprecated, non-zero values.
-	ValuesForExamples []*SampleValue
-	// Parent returns the ancestor of this node, if any.
-	Parent *Message
-	// The Protobuf package this enum belongs to.
-	Package string
-	// Language specific annotations.
-	Codec any
-}
-
-// EnumValue defines a value in an Enum.
-type EnumValue struct {
-	// Documentation for the message.
-	Documentation string
-	// Name of the attribute.
-	Name string
-	// ID is a unique identifier.
-	ID string
-	// Some source specifications allow marking enum values as deprecated.
-	Deprecated bool
-	// Number of the attribute.
-	Number int32
-	// Parent returns the ancestor of this node, if any.
-	Parent *Enum
-	// Language specific annotations.
-	Codec any
-}
-
-// SampleValue represents a value used in a sample.
-type SampleValue struct {
-	// The enum value.
-	EnumValue *EnumValue
-	// The index of the value in the sample list (0-based).
-	Index int
-}
-
-// Field defines a field in a Message.
-type Field struct {
-	// Documentation for the field.
-	Documentation string
-	// Name of the attribute.
-	Name string
-	// ID is a unique identifier.
-	ID string
-	// Typez is the datatype of the field.
-	Typez Typez
-	// TypezID is the ID of the type the field refers to. This value is populated
-	// for message-like types only.
-	TypezID string
-	// JSONName is the name of the field as it appears in JSON. Useful for
-	// serializing to JSON.
-	JSONName string
-	// Optional indicates that the field is marked as optional in proto3.
-	Optional bool
-
-	// For a given field, at most one of `Repeated` or `Map` is true.
-	//
-	// Using booleans (as opposed to an enum) makes it easier to write mustache
-	// templates.
-	//
-	// Repeated is true if the field is a repeated field.
-	Repeated bool
-	// Map is true if the field is a map.
-	Map bool
-	// Some source specifications allow marking fields as deprecated.
-	Deprecated bool
-	// IsOneOf is true if the field is related to a one-of and not
-	// a proto3 optional field.
-	IsOneOf bool
-	// Some fields have a type that refers (sometimes indirectly) to the
-	// containing message. That triggers slightly different code generation for
-	// some languages.
-	Recursive bool
-	// AutoPopulated is true if the field is eligible to be auto-populated,
-	// per the requirements in AIP-4235.
-	//
-	// That is:
-	// - It has Typez == TypezString
-	// - For Protobuf, does not have the `google.api.field_behavior = REQUIRED` annotation
-	// - For Protobuf, has the `google.api.field_info.format = UUID4` annotation
-	// - For OpenAPI, it is an optional field
-	// - For OpenAPI, it has format == "uuid"
-	AutoPopulated bool
-	// FieldBehavior indicates how the field behaves in requests and responses.
-	//
-	// For example, that a field is required in requests, or given as output
-	// but ignored as input.
-	Behavior []FieldBehavior
-	// For fields that are part of a OneOf, the group of fields that makes the
-	// OneOf.
-	Group *OneOf
-	// The message that contains this field.
-	Parent *Message
-	// The message type for this field, can be nil.
-	MessageType *Message
-	// The enum type for this field, can be nil.
-	EnumType *Enum
-	// ResourceReference contains the data from the `google.api.resource_reference`
-	// annotation.
-	ResourceReference *ResourceReference
-	// ResourceNamePattern is a parsed representation of the resource pattern associated
-	// with this field.
-	ResourceNamePattern *ResourceNamePattern
-	// Codec is a placeholder to put language specific annotations.
-	Codec any
-}
-
-// DocumentAsRequired returns true if the field should be documented as required.
-func (field *Field) DocumentAsRequired() bool {
-	return slices.Contains(field.Behavior, FieldBehaviorRequired)
-}
-
-// Singular returns true if the field is not a map or a repeated field.
-func (f *Field) Singular() bool {
-	return !f.Map && !f.Repeated
-}
-
-// NameEqualJSONName returns true if the field's name is the same as its JSON name.
-func (f *Field) NameEqualJSONName() bool {
-	return f.JSONName == f.Name
-}
-
-// IsString returns true if the primitive type of a field is `TypezString`.
-//
-// This is useful for mustache templates that differ only
-// in the broad category of field type involved.
-func (f *Field) IsString() bool {
-	return f.Typez == TypezString
-}
-
-// IsBytes returns true if the primitive type of a field is `TypezBytes`.
-//
-// This is useful for mustache templates that differ only
-// in the broad category of field type involved.
-func (f *Field) IsBytes() bool {
-	return f.Typez == TypezBytes
-}
-
-// IsBool returns true if the primitive type of a field is `TypezBool`.
-//
-// This is useful for mustache templates that differ only
-// in the broad category of field type involved.
-func (f *Field) IsBool() bool {
-	return f.Typez == TypezBool
-}
-
-// IsLikeInt returns true if the primitive type of a field is one of the
-// integer types.
-//
-// This is useful for mustache templates that differ only
-// in the broad category of field type involved.
-func (f *Field) IsLikeInt() bool {
-	switch f.Typez {
-	case TypezInt32, TypezInt64, TypezSint32, TypezSint64:
-		return true
-	case TypezSfixed32, TypezSfixed64:
-		return true
-	default:
-		return false
-	}
-}
-
-// IsLikeUInt returns true if the primitive type of a field is one of the
-// unsigned integer types.
-//
-// This is useful for mustache templates that differ only
-// in the broad category of field type involved.
-func (f *Field) IsLikeUInt() bool {
-	switch f.Typez {
-	case TypezUint32, TypezUint64, TypezFixed32, TypezFixed64:
-		return true
-	default:
-		return false
-	}
-}
-
-// IsLikeFloat returns true if the primitive type of a field is a float or
-// double.
-//
-// This is useful for mustache templates that differ only
-// in the broad category of field type involved.
-func (f *Field) IsLikeFloat() bool {
-	return f.Typez == TypezDouble || f.Typez == TypezFloat
-}
-
-// IsEnum returns true if the primitive type of a field is `TypezEnum`.
-//
-// This is useful for mustache templates that differ only
-// in the broad category of field type involved.
-func (f *Field) IsEnum() bool {
-	return f.Typez == TypezEnum
-}
-
-// IsObject returns true if the primitive type of a field is `OBJECT_TYPE`.
-//
-// This is useful for mustache templates that differ only
-// in the broad category of field type involved.
-//
-// The templates *should* first check if the field is singular, as all maps are
-// also objects.
-func (f *Field) IsObject() bool {
-	return f.Typez == TypezMessage
-}
-
-// OneOf is a group of fields that are mutually exclusive. Notably, proto3 optional
-// fields are all their own one-of.
-type OneOf struct {
-	// Name of the attribute.
-	Name string
-	// ID is a unique identifier.
-	ID string
-	// Documentation for the field.
-	Documentation string
-	// Fields associated with the one-of.
-	Fields []*Field
-	// The best field to show in a oneof related samples.
-	// Non deprecated fields are preferred, then scalar, repeated, map fields
-	// in that order.
-	ExampleField *Field
-	// Codec is a placeholder to put language specific annotations.
-	Codec any
-}
-
-// Resource is a fundamental building block of an API, representing an
-// individually-named entity (a "noun").
-//
-// Resources are typically organized into a hierarchy, where each node is either a simple resource or a
-// collection of resources.
-// This definition is based on AIP-121 (https://google.aip.dev/121).
-type Resource struct {
-	// Type identifies the kind of resource (e.g., "cloudresourcemanager.googleapis.com/Project").
-	// This string is globally unique and identifies the type of resource across Google Cloud.
-	Type string
-	// Pattern is a list of resource patterns, where each pattern is a sequence of path segments.
-	// This defines the structure of the resource's unique identifier.
-	Patterns []ResourcePattern
-	// Plural is the plural form of the resource name.
-	// For example, for a "Book" resource, Plural would be "books".
-	Plural string
-	// Singular is the singular form of the resource name.
-	// For example, for a "Book" resource, Singular would be "book".
-	Singular string
-	// Self points to the Message that defines this resource.
-	// This creates a back-reference for navigating the API model,
-	// allowing a Resource definition to access its originating Message structure.
-	Self *Message
-	// Language specific annotations.
-	Codec any
-}
-
-// ResourcePattern is a sequence of path segments that defines the structure of a resource's unique identifier.
-//
-// Given a resource name pattern like `projects/{project}/locations/{region}/secret/{secret}` this
-// will be:
-//
-//	[]PathSegment{
-//	  {Literal: "projects"},
-//	  {Variable: "project"},
-//	  {Literal: "locations"},
-//	  {Variable: "region"},
-//	  {Literal: "secrets"},
-//	  {Variable: "secret"},
-//	}
-type ResourcePattern []PathSegment
-
-// ResourceNameSegment is a segment of a resource name pattern.
-//
-// This should be a union type, either `Literal` or `Variable` are set, but not both.
-type ResourceNameSegment struct {
-	// Literal is the literal part of the segment.
-	//
-	// When formatting a resource name this should be used, well, literally.
-	Literal string
-
-	// Variable is the name of the variable part of the segment.
-	//
-	// When formatting a resource name, this names a variable used to populate the resource name.
-	Variable string
-}
-
-// ResourceNamePattern describes the structure of a resource name.
-type ResourceNamePattern struct {
-	Segments []ResourceNameSegment
-}
-
-// ResourceReference describes a field's relationship to another resource type.
-// It acts as a foreign key, indicating that the field's value identifies an instance of another resource.
-// This relationship is established via the `google.api.resource_reference` annotation in Protobuf.
-type ResourceReference struct {
-	// Type is the unique identifier of the referenced resource's kind (e.g., "library.googleapis.com/Shelf").
-	// This string matches the `Type` field in the corresponding `Resource` definition.
-	Type string
-	// ChildType is the unique identifier of a *child* resource's kind.
-	// This is used when a field references a parent resource (e.g., "Shelf"), but the context
-	// implies interaction with a specific child type (e.g., "Book" within that shelf).
-	ChildType string
-	// Language specific annotations.
-	Codec any
-}
-
-// IsResourceReference returns true if the field is annotated with google.api.resource_reference.
-func (f *Field) IsResourceReference() bool {
-	return f.ResourceReference != nil
-}
-
-// TargetResource contains the results of the resource name identification.
-// It provides the sequences of fields used by language-specific generators to inject tracing attributes.
-type TargetResource struct {
-	// FieldPaths is a list of field name sequences that, when joined, form a resource name.
-	// For example, [["project"], ["zone"], ["instance"]] identifies a multi-part resource.
-	FieldPaths [][]string
-
-	// Template is the canonical HTTP path template for the resource, derived from the PathBinding's PathTemplate by removing the API version prefix.
-	// For example, if the PathTemplate is "//compute.googleapis.com/projects/{project}/zones/{zone}", the Template will be a []PathSegment containing:
-	// - a Literal segment for "//compute.googleapis.com"
-	// - a Literal segment for "projects"
-	// - a Variable segment with FieldPath ["project"]
-	// - a Literal segment for "zones"
-	// - a Variable segment with FieldPath ["zone"]
-	Template []PathSegment
-}
