+++ internal/sidekick/api/operation_info.go
@@ -1,0 +1,29 @@
+// Copyright 2026 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package api
+
+// OperationInfo contains normalized long running operation info.
+type OperationInfo struct {
+	// The metadata type. If there is no metadata, this is set to
+	// `.google.protobuf.Empty`.
+	MetadataTypeID string
+	// The result type. This is the expected type when the long running
+	// operation completes successfully.
+	ResponseTypeID string
+	// The method.
+	Method *Method
+	// Language specific annotations.
+	Codec any
+}
