+++ internal/sidekick/api/path_binding.go
@@ -1,0 +1,37 @@
+// Copyright 2026 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package api
+
+// PathBinding is a binding of a path to a method.
+type PathBinding struct {
+	// HTTP Verb.
+	//
+	// This is one of:
+	// - GET
+	// - POST
+	// - PUT
+	// - DELETE
+	// - PATCH
+	Verb string
+	// The path broken by components.
+	PathTemplate *PathTemplate
+	// Query parameter fields.
+	QueryParameters map[string]bool
+	// TargetResource contains the results of the resource name identification.
+	// This helps identify which resource this path is likely targeting.
+	TargetResource *TargetResource
+	// Language specific annotations.
+	Codec any
+}
